refactor: 커스텀 훅 분리 및 상태관리 최소화
