+++ src/pages/community/hooks/index.ts
@@ -1,3 +1,4 @@
 import useModal from './useModal';
+import useToolCategorySelect from './useToolCategorySelect';
 
-export { useModal };
+export { useModal, useToolCategorySelect };
